requirements.txt
Package                   Version
------------------------- -----------
absl-py                   2.3.0
altair                    5.5.0
attrs                     25.3.0
blinker                   1.9.0
cachetools                5.5.2
certifi                   2025.4.26
cffi                      1.17.1
charset-normalizer        3.4.2
click                     8.2.1
colorama                  0.4.6
contourpy                 1.3.2
cycler                    0.12.1
flatbuffers               25.2.10
fonttools                 4.58.1
gitdb                     4.0.12
GitPython                 3.1.44
idna                      3.10
jax                       0.6.1
jaxlib                    0.6.1
Jinja2                    3.1.6
jsonschema                4.24.0
jsonschema-specifications 2025.4.1
kiwisolver                1.4.8
MarkupSafe                3.0.2
matplotlib                3.10.3
mediapipe                 0.10.21
ml_dtypes                 0.5.1
narwhals                  1.41.0
numpy                     1.26.4
opencv-python             4.11.0.86
opencv-python-headless    4.11.0.86
opt_einsum                3.4.0
packaging                 24.2
pandas                    2.3.0
pillow                    11.2.1
pip                       25.1.1
protobuf                  4.25.8
pyarrow                   20.0.0
pycparser                 2.22
pydeck                    0.9.1
pyparsing                 3.2.3
python-dateutil           2.9.0.post0
pytz                      2025.2
referencing               0.36.2
requests                  2.32.3
rpds-py                   0.25.1
scipy                     1.15.3
sentencepiece             0.2.0
setuptools                57.4.0
six                       1.17.0
smmap                     5.0.2
sounddevice               0.5.2
streamlit                 1.45.1
tenacity                  9.1.2
toml                      0.10.2
tornado                   6.5.1
typing_extensions         4.14.0
tzdata                    2025.2
urllib3                   2.4.0
watchdog                  6.0.0
